+++ app/twitter-image.tsx
@@ -1,10 +1,12 @@
 import { ImageResponse } from "next/server";
+import { allPosts } from "contentlayer/generated";
+import { format, parseISO } from "date-fns";
 
 import { defaultAuthor } from "@/lib/metadata";
 
 export const runtime = "edge";
 
-export const alt = `About ${defaultAuthor.name}`;
+export const alt = `Article by ${defaultAuthor.name}`;
 export const size = {
   width: 1200,
   height: 630,
@@ -13,22 +15,69 @@ export const size = {
 export const contentType = "image/png";
 
 // Image generation
-export default async function Image() {
+export default async function Image({ params }: { params: { slug: string } }) {
+  const post = await allPosts.find((post) => post.slug === params.slug);
+
+  if (!post) {
+    return {};
+  }
+
+  const date = post.lastUpdatedDate || post.publishedDate;
+
   return new ImageResponse(
     (
       // ImageResponse JSX element
       <div
         style={{
-          fontSize: 128,
-          background: "white",
-          width: "100%",
-          height: "100%",
+          background: "linear-gradient(45deg, rgba(59, 178, 93, 0.20) 0%, rgba(59, 121, 178, 0.20) 100%)",
           display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
+          height: "100%",
+          width: "100%",
+          alignItems: "flex-start",
+          flexDirection: "column",
+          justifyContent: "space-between",
+          letterSpacing: "-.02em",
+          padding: "64px 48px",
+          color: "#222",
         }}
       >
-        {`${defaultAuthor.handle}'s blog`}
+        <div style={{ display: "flex" }}>
+          <span
+            style={{
+              fontSize: "24px",
+              fontWeight: 400,
+            }}
+          >
+            {defaultAuthor.handle}
+          </span>
+        </div>
+        <div
+          style={{
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "flex-start",
+            width: "auto",
+            maxWidth: "70%",
+          }}
+        >
+          <p
+            style={{
+              fontWeight: "bold",
+              fontSize: "48px",
+              lineHeight: 1.1,
+            }}
+          >
+            {post.title}
+          </p>
+
+          <p
+            style={{
+              fontSize: "20px",
+            }}
+          >
+            {format(parseISO(date), "LLLL d, yyyy")} &middot; {post.readTimeMinutes} min read
+          </p>
+        </div>
       </div>
     ),
     // ImageResponse options